Organize entities, scene and map into different folders
