+++ src/entities/index.ts
@@ -1,0 +1,10 @@
+const entities = {
+  player: {
+    1: null,
+  },
+  bullets: null as Phaser.GameObjects.Group | null,
+  enemyGroup: null as Phaser.GameObjects.Group | null,
+  turretGroup: null as Phaser.GameObjects.Group | null,
+};
+
+export default entities;
