+++ src/map/constants.ts
@@ -1,0 +1,22 @@
+import {
+  ORIENTATION_VERTICAL,
+  ORIENTATION_HORIZONTAL
+} from '../constants';
+
+export const PATH_SEGMENTS = [
+  {
+    orientation: ORIENTATION_VERTICAL,
+    start: { x: 2, y: 0 },
+    size: 6,
+  },
+  {
+    orientation: ORIENTATION_HORIZONTAL,
+    start: { x: 2, y: 5 },
+    size: 13,
+  },
+  {
+    orientation: ORIENTATION_VERTICAL,
+    start: { x: 14, y: 5 },
+    size: 11,
+  },
+];
