+++ src/utils.ts
@@ -2,3 +2,5 @@
 export const noop = () => {};
 
 export const clone = (b) => JSON.parse(JSON.stringify(b));
+
+export const isDebugMode = !(window.location.href.indexOf('preview=true') != -1);
